+++ R/h5.R
@@ -180,3 +180,21 @@ write_h5_ <- function(x, file_h5, path) {
   }
 }
 
+
+
+
+#' Check if a matrix is sparse
+#'
+#' @param x A sparse or not sparse matrix
+#'
+#' @export
+#'
+#' @examples
+#' is_sparse(matrix(c(1:10)))
+#' is_sparse(Matrix::Matrix(matrix(c(1:10))))
+is_sparse <- function(x) {
+  any(grepl("[di]..Matrix", class(x)))
+}
+
+
+